refactor: remove postfix '_util' from python scripts inside util dir
