+++ srpcLib/stub_generator/python/server_python_stub_generator.py
@@ -2,7 +2,7 @@ import logging
 import textwrap
 
 from ...interface.srpc_stub_generator_interface import SrpcStubGeneratorInterface
-from ...utils.srpc_stub_util import (
+from ...utils.srpc_stub import (
     DEFAULT_CONNECTION_PORT,
     LIB_NAME,
     get_service_interface_class_name,
@@ -23,7 +23,7 @@ import inspect
 import os
 
 from {LIB_NAME}.utils.srpc_serializer import SrpcSerializer
-import {LIB_NAME}.utils.srpc_network_util as srpcnetwork
+import {LIB_NAME}.utils.srpc_network as srpcnetwork
 import logging
 
 """
